examples/crewai-flow-validation/requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
httpx>=0.24.0
python-dotenv>=1.0.0
